+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ New features
 * Add :code:`modeling.add_arm`
 * Add :code:`modeling.add_observations`
 * Add :code:`modeling.add_administration`
+* Add :code:`modeling.create_dataset_from_design`
 
 Bugfixes
 ========